* fix precision of after-tax price in BO

+++ backtoshops/sales/templatetags/calc_discount_price.py
@@ -10,7 +10,7 @@ def calc_discount_price(discount_type, discount, base_price):
         base_price = float(base_price)
         if discount_type == "percentage":
             tmp = base_price - (base_price * (discount / 100))
-            return round(int(tmp * 100) + ((tmp * 1000 % 10 > 4) and 1 or 0)) / 100
+            return round(int(tmp * 100) + (int(tmp * 1000 % 10) > 4 and 1 or 0)) / 100
         else:
             return base_price - discount
     except Exception, e:
@@ -22,7 +22,7 @@ def calc_discount_price(discount_type, discount, base_price):
 @register.simple_tag
 def calc_price_after_tax(price, tax_rate):
     tmp = price + float(price * tax_rate / 100.0)
-    return round(int(tmp * 100) + ((tmp * 1000 % 10 > 4) and 1 or 0)) / 100
+    return round(int(tmp * 100) + (int(tmp * 1000 % 10) > 4 and 1 or 0)) / 100
 
 @register.simple_tag
 def calc_discount_price_after_tax(discount_type, discount, base_price, tax_rate):
